+++ src/components/Portal.js
@@ -1,5 +1,6 @@
-import React, { Component, PropTypes } from 'react';
-import Transition from 'react-addons-css-transition-group';
+import PropTypes from 'prop-types';
+import React, { Component } from 'react';
+import CSSTransitionGroup from 'react-transition-group/CSSTransitionGroup';
 import { render } from 'react-dom';
 import PassContext from './PassContext';
 
@@ -29,7 +30,7 @@ export default class Portal extends Component {
 			<PassContext context={this.context}>
 				<div>
 					<style>{styles}</style>
-					<Transition
+					<CSSTransitionGroup
 						component="div"
 						transitionName="fade"
 						transitionEnterTimeout={duration}