+++ src/components/Thumbnail.js
@@ -1,4 +1,5 @@
-import React, { PropTypes } from 'react';
+import PropTypes from 'prop-types';
+import React from 'react';
 import { css, StyleSheet } from 'aphrodite/no-important';
 
 import defaults from '../theme';
@@ -12,10 +13,9 @@ function Thumbnail ({ index, src, thumbnail, active, onClick }, { theme }) {
 		<div
 			className={css(classes.thumbnail, active && classes.thumbnail__active)}
 			onClick={(e) => {
-				e.preventDefault()
-				e.stopPropagation()
-				
-				onClick(index)
+				e.preventDefault();
+				e.stopPropagation();
+				onClick(index);
 			}}
 			style={{ backgroundImage: 'url("' + url + '")' }}
 		/>